+++ RELEASE_NOTES.md
@@ -52,6 +52,13 @@ set to `warning`.
 * FIXED: When expanding a file name, it's only necessary to ensure that its
 directory exists for a generated intermediate file.
 
+Prolog adapter and integration files
+------------------------------------
+
+* IMPROVED: The adapter files for SICStus Prolog, SWI-Prolog, and YAP now
+define the `message_hook/3` predicate to avoid printing repeated discontiguous
+predicate warnings that are already reported by the Logtalk compiler.
+
 Documentation
 -------------